fix: use u64 for VM addresses in Section enum to support 32-bit hosts
The Section enum was using `usize` for virtual addresses, which caused
truncation on 32-bit platforms (e.g., RISC-V zkVM). VM addresses like
MM_RODATA_START (0x100000000) would truncate to 0 when cast to 32-bit
usize, causing "Access violation in program section" errors.

This change distinguishes between:
- VM addresses (always u64 per sBPF spec)
- Host addresses (native usize for memory pointers)

Changes:
- Section::Owned and Section::Borrowed now use u64 for VM addresses
- parse_ro_sections() uses u64 for lowest_addr, highest_addr, ro_slices
- Removed truncating `as usize` casts for MM_RODATA_START
- Updated test assertions to compare map() results with `as u64`

+++ src/memory_region.rs
@@ -821,12 +821,12 @@ mod test {
 
         assert_eq!(
             m.map(AccessType::Load, ebpf::MM_INPUT_START, 1).unwrap(),
-            mem1.as_ptr() as usize
+            mem1.as_ptr() as u64
         );
 
         assert_eq!(
             m.map(AccessType::Store, ebpf::MM_INPUT_START, 1).unwrap(),
-            mem1.as_ptr() as usize
+            mem1.as_ptr() as u64
         );
 
         assert_error!(
@@ -841,7 +841,7 @@ mod test {
                 1,
             )
             .unwrap(),
-            mem2.as_ptr() as usize
+            mem2.as_ptr() as u64
         );
 
         assert_eq!(
@@ -851,7 +851,7 @@ mod test {
                 1,
             )
             .unwrap(),
-            mem3.as_ptr() as usize
+            mem3.as_ptr() as u64
         );
 
         assert_eq!(
@@ -861,7 +861,7 @@ mod test {
                 1,
             )
             .unwrap(),
-            mem4.as_ptr() as usize
+            mem4.as_ptr() as u64
         );
 
         assert_error!(
@@ -1117,7 +1117,7 @@ mod test {
 
         assert_eq!(
             m.map(AccessType::Load, ebpf::MM_INPUT_START, 1).unwrap(),
-            mem1.as_ptr() as usize
+            mem1.as_ptr() as u64
         );
 
         assert_eq!(
@@ -1127,7 +1127,7 @@ mod test {
                 1,
             )
             .unwrap(),
-            mem2.as_ptr() as usize
+            mem2.as_ptr() as u64
         );
 
         assert_error!(
@@ -1167,7 +1167,7 @@ mod test {
                 1,
             )
             .unwrap(),
-            mem3.as_ptr() as usize
+            mem3.as_ptr() as u64
         );
     }
 
@@ -1192,7 +1192,7 @@ mod test {
 
         assert_eq!(
             m.map(AccessType::Load, ebpf::MM_STACK_START, 1).unwrap(),
-            mem2.as_ptr() as usize
+            mem2.as_ptr() as u64
         );
 
         // index > regions.len()
@@ -1221,7 +1221,7 @@ mod test {
 
         assert_eq!(
             m.map(AccessType::Load, ebpf::MM_STACK_START, 1).unwrap(),
-            mem3.as_ptr() as usize
+            mem3.as_ptr() as u64
         );
     }
 
@@ -1253,12 +1253,12 @@ mod test {
             assert_eq!(
                 m.map_with_access_violation_handler(AccessType::Load, ebpf::MM_RODATA_START, 1)
                     .unwrap(),
-                original.as_ptr() as usize
+                original.as_ptr() as u64
             );
             assert_eq!(
                 m.map_with_access_violation_handler(AccessType::Store, ebpf::MM_RODATA_START, 1)
                     .unwrap(),
-                copied.borrow().as_ptr() as usize
+                copied.borrow().as_ptr() as u64
             );
         }
     }
@@ -1290,7 +1290,7 @@ mod test {
 
             assert_eq!(
                 m.map(AccessType::Load, ebpf::MM_RODATA_START, 1).unwrap(),
-                original.as_ptr() as usize
+                original.as_ptr() as u64
             );
 
             assert_eq!(m.load::<u8>(ebpf::MM_RODATA_START).unwrap(), 11);
